refactor(chat): message list + message subcomponents
