+++ frontend/src/routes/chat/[id]/+page.svelte
@@ -1,15 +1,13 @@
 <script lang="ts">
-  import type { ChatMessage, ChatState, ChatSource } from "$lib/types/chat";
-  import { scrollToPosition, handleScroll, initScrollCheck, positionTooltip, hideTooltip } from "$lib/utils/sources";
-  import { processReasoningMessage, processOutputMessage, toggleReasoning, processSource } from '$lib/utils/chat';
-  import { renderLatex, renderContent } from '$lib/utils/katex';
+  import type { ChatMessage, ChatSource, ChatState } from "$lib/types/chat";
   import ChatInput from '$lib/components/chat/chat-input.svelte';
+  import MessageList from '$lib/components/chat/message-list.svelte';
   import "katex/dist/katex.min.css";
   
-  import { CheckIcon, ChevronDownIcon, FileIcon, FileTextIcon, HardDriveIcon, SendIcon } from "svelte-feather-icons";
   import { onMount } from "svelte";
   import { conversationStore } from "$lib/stores/conversationStore";
   import { modelStore } from "$lib/stores/modelStore";
+  import { processOutputMessage, processReasoningMessage, processSource } from "$lib/utils/chat";
 
   export let data: { 
     id: string, 
@@ -20,9 +18,7 @@
   };
   
   let messages: ChatMessage[] = [];
-  let conversationContainer: HTMLDivElement;
   let isReasoning = false;
-  let truncateLength = 80;
   let isLoading = false;
   let eventSource: EventSource | null = null;
   let currentConversationId: string | null = null;
@@ -59,6 +55,12 @@
         messages = data.conversation;
       }
     }
+
+    return () => {
+      eventSource?.close();
+      eventSource = null;
+      isLoading = false;
+    };
   });
   
   async function handleSendMessage(event: { detail: { query: string } }) {
@@ -122,7 +124,7 @@
     const url = `/chat?conversationId=${encodeURIComponent(data.id)}`;
     eventSource = new EventSource(url);
     let botMessage : ChatMessage;
-    
+
     // Reference the last message in the messages array if it's a bot message
     // This ensures we're updating the correct message when streaming
     if (messages.length > 0 && messages[messages.length - 1].sender === 'bot') {
@@ -219,185 +221,7 @@
 
 <main class="min-h-[calc(100vh-60px)] flex flex-col items-center justify-center px-6">
   <div class="flex-1 flex flex-col w-full max-w-3xl h-screen">
-    <div class="conversation-container flex-1 overflow-y-auto !overflow-x-hidden p-4 space-y-6 pb-32 max-w-full w-full" style="height: calc(100vh - 100px);">
-      {#each messages as message, messageIndex}
-        <div class="space-y-4">
-          <div class="prose max-w-3xl mx-auto prose-lg w-full overflow-x-hidden">
-            {#if message.sender === 'user'}
-              <div class="font-bold prose-xl break-words whitespace-normal overflow-hidden w-full">{message.text}</div>
-            {:else}
-              {#if message.sources.length > 0}
-                <div class="mb-6">
-                    <div class="flex justify-between items-center mb-2">
-                        <h3 class="text-sm font-semibold text-gray-600">Sources</h3>
-                        {#if message.isScrollable}
-                        <button 
-                            class="text-xs text-gray-400 hover:text-gray-600 transition-colors cursor-pointer"
-                            on:click={(e) => {
-                                const target = e.target as HTMLElement;
-                                const scrollContainer = target.closest('.mb-6')?.querySelector('.scroll-container');
-                                if (scrollContainer) {
-                                    scrollToPosition(messages, scrollContainer as HTMLElement, messageIndex, (msgs) => {
-                                        messages = msgs;
-                                    });
-                                }
-                            }}
-                        >
-                            {message.sourcesScrollAtEnd ? '← Scroll to start' : 'Scroll to end →'}
-                        </button>
-                        {/if}
-                    </div>
-
-                    <!-- Sources -->
-                    <div class="relative !overflow-hidden w-full">
-                        <div 
-                            class="flex overflow-x-auto overflow-y-hidden pb-4 gap-3 scrollbar-thin scroll-container max-w-full"
-                            on:scroll={(e) => {
-                                messages = handleScroll(messages, e, messageIndex);
-                            }}
-                            use:initScrollCheck={{
-                              node: document.querySelector('.conversation-container'), 
-                              messages, 
-                              messageIndex,
-                              setMessages: (msgs) => {
-                                  messages = msgs;
-                              }
-                            }}
-                        >
-                            {#each message.sources as source, sourceIndex}
-                                <div class="flex-none w-[325px]">
-                                    <div 
-                                        class="bg-white rounded-xl p-3 hover:bg-gray-100 transition-colors duration-200 shadow-sm border border-gray-100 relative tooltip-container"
-                                        on:mouseenter={positionTooltip}
-                                        on:mouseleave={hideTooltip}
-                                        data-tooltip-id={`tooltip-${messageIndex}-${sourceIndex}`}
-                                        role="button"
-                                        tabindex="0"
-                                        aria-describedby={`tooltip-${messageIndex}-${sourceIndex}`}
-                                    >
-                                        <div class="flex items-center gap-1 text-gray-400 mb-1">
-                                            <HardDriveIcon size="14" />
-                                            <span class="text-gray-300 mx-1">|</span>
-                                            {#if source.extension === 'pdf'}
-                                                <FileTextIcon size="14" />
-                                            {:else}
-                                                <FileIcon size="14" />
-                                            {/if}
-                                            <span class="text-xs uppercase tracking-wider font-medium">
-                                                {source.extension}
-                                            </span>
-                                        </div>
-                                        <div>
-                                            <div class="text-sm font-medium text-gray-900 truncate mb-1">{source.title}</div>
-                                            <div class="text-xs text-gray-500 truncate font-light">{source.filePath}</div>
-                                        </div>
-                                        
-                                        <!-- Source tooltip that appears on hover -->
-                                        <div class="tooltip fixed bg-white opacity-0 pointer-events-none text-gray-800 p-3 rounded-xl shadow-md text-sm z-20 max-w-full whitespace-normal border border-gray-100" 
-                                             id={`tooltip-${messageIndex}-${sourceIndex}`}
-                                             role="tooltip"
-                                             aria-hidden="true">
-
-                                            <div class="font-semibold mb-1 break-words">{source.title}</div>
-                                            <div class="text-xs text-gray-600 break-words">{source.filePath}</div>
-                                        </div>
-                                    </div>
-                                </div>
-                            {/each}
-                        </div>
-                    </div>
-                </div>
-                
-              {/if}
-              {#if message.reasoning.length > 0}
-                <div class="max-w-3xl mx-auto">
-                  <div class="flex justify-between items-center">
-                    <h3 class="text-sm font-semibold text-gray-600">Reasoning</h3>
-                    <button
-                      class="text-gray-600 cursor-pointer transition-transform duration-200 mt-3"
-                      class:rotate-180={!message.reasoningSectionCollapsed}
-                      on:click={() => {
-                        if (message.sender !== 'user') {
-                          messages = messages.map((msg, idx) =>
-                            idx === messageIndex
-                              ? {
-                                  ...msg,
-                                  reasoningSectionCollapsed: !msg.reasoningSectionCollapsed
-                                }
-                              : msg
-                          );
-                        }
-                      }}
-                    >
-                      <ChevronDownIcon size="16" />
-                    </button>
-                  </div>
-
-                  {#if !message.reasoningSectionCollapsed}
-                    {#each message.reasoning as thought, reasoningIndex}
-                      <div class="pl-3 py-2 mb-3 w-full">
-                        <div class="flex items-start w-full">
-                          <div class="flex justify-between items-start gap-2 w-full">
-                            <div class="flex items-start gap-2 flex-1 min-w-0">
-                              <div class="shrink-0">
-                                {#if isReasoning && reasoningIndex === message.reasoning.length - 1}
-                                  <div class="relative mt-3">
-                                    <div class="w-2 h-2 bg-green-400 rounded-full"></div>
-                                    <div
-                                      class="w-2 h-2 bg-green-400 rounded-full absolute top-0 animate-ping"
-                                    ></div>
-                                  </div>
-                                {:else if isReasoning}
-                                  <div class="w-2 h-2 bg-gray-400 rounded-full mt-3"></div>
-                                {:else}
-                                  <CheckIcon size="16" class="text-gray-500 mt-2" />
-                                {/if}
-                              </div>
-                              <div class="text-gray-600 reasoning-container">
-                                <div
-                                  class={`reasoning-content ${thought.collapsed ? 'collapsed' : 'expanded'}`}
-                                >
-                                  {@html renderLatex(thought.text)}
-                                </div>
-                              </div>
-                            </div>
-                            {#if thought.text.length > truncateLength}
-                              <button
-                                class="text-gray-600 shrink-0 cursor-pointer transition-transform duration-200 mt-2"
-                                class:rotate-180={!thought.collapsed}
-                                on:click={() => {
-                                  const state = {
-                                    messages,
-                                    isReasoning
-                                  };
-                                  toggleReasoning(messageIndex, reasoningIndex, state);
-                                  messages = state.messages;
-                                }}
-                              >
-                                <ChevronDownIcon size="16" />
-                              </button>
-                            {/if}
-                          </div>
-                        </div>
-                      </div>
-                    {/each}
-                  {/if}
-                </div>
-              {/if}
-              {#if message.text !== ''}
-                <h3 class="text-sm font-semibold text-gray-600">Answer</h3>
-                <div class="mt-4 prose max-w-3xl mx-auto prose-lg">
-                  {@html renderContent(message.text)}
-                </div>
-              {:else}
-                <div class="animate-pulse mt-4">Thinking...</div>
-              {/if}
-            {/if}
-          </div>
-        </div>
-      {/each}
-    </div>
-    <!-- Chat Input -->
+    <MessageList {messages} {isReasoning} />
     <ChatInput 
       on:send={handleSendMessage} 
       {isLoading} 
@@ -411,166 +235,5 @@
     overflow-x: hidden;
     position: relative;
     width: 100%;
-  }
-  
-  .reasoning-container {
-    position: relative;
-    width: 100%;
-    overflow: hidden;
-  }
-
-  .reasoning-content {  
-    overflow: hidden;
-  }
-
-  .reasoning-content.collapsed {
-    max-height: 1.5em;
-    text-overflow: ellipsis;
-    white-space: nowrap;
-    overflow: hidden;
-  }
-
-  .reasoning-content.expanded {
-    max-height: 500px;
-    white-space: normal;
-    overflow: hidden;
-  }
-
-  .scrollbar-thin {
-      scrollbar-width: thin;
-      -ms-overflow-style: none;
-      scroll-behavior: smooth;
-  }
-
-  .scrollbar-thin::-webkit-scrollbar {
-      height: 6px;
-  }
-
-  .scrollbar-thin::-webkit-scrollbar-track {
-      background: #f1f1f1;
-      border-radius: 3px;
-      margin: 0;
-  }
-
-  .scrollbar-thin::-webkit-scrollbar-thumb {
-      background: #888;
-      border-radius: 3px;
-  }
-
-  .scrollbar-thin::-webkit-scrollbar-thumb:hover {
-      background: #666;
-  }
-  
-  /* No scrollbar class to hide scrollbars but maintain functionality */
-  .no-scrollbar {
-      -ms-overflow-style: none;  /* IE and Edge */
-      scrollbar-width: none;  /* Firefox */
-  }
-  
-  .no-scrollbar::-webkit-scrollbar {
-      display: none;  /* Chrome, Safari and Opera */
-  }
-
-  .group {
-      position: relative;
-  }
-
-  /* Prevent tooltip from being cut off */
-  .scroll-container {
-      margin-top: 0;
-      padding-top: 0;
-  }
-
-  /* Remove previous tooltip styles and add these */
-  .pointer-events-none {
-      pointer-events: none;
-  }
-  
-  .pointer-events-auto {
-      pointer-events: auto;
-  }
-
-  /* Add tooltip styles */
-  .tooltip-container {
-      position: relative;
-  }
-  
-  .tooltip {
-      width: 325px;
-      box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.1), 0 2px 4px -1px rgba(0, 0, 0, 0.06);
-      transition: opacity 0.2s, visibility 0.2s, transform 0.2s;
-      transition-delay: 300ms;
-      visibility: hidden;
-      position: fixed;
-      z-index: 50;
-      transform: scaleY(0.98);
-      transform-origin: top center;
-  }
-  
-  .tooltip-container:hover .tooltip {
-      opacity: 1;
-      visibility: visible;
-      transform: scaleY(1);
-  }
-
-  /* Textarea scrollbar styling */
-  .textarea-scrollbar {
-    scrollbar-width: thin;
-    -ms-overflow-style: none;
-    scroll-behavior: smooth;
-  }
-
-  .textarea-scrollbar::-webkit-scrollbar {
-    width: 6px;
-  }
-
-  .textarea-scrollbar::-webkit-scrollbar-track {
-    background: #f1f1f1;
-    border-radius: 3px;
-    margin: 0;
-  }
-
-  .textarea-scrollbar::-webkit-scrollbar-thumb {
-    background: #888;
-    border-radius: 3px;
-  }
-
-  .textarea-scrollbar::-webkit-scrollbar-thumb:hover {
-    background: #666;
-  }
-
-  .textarea-container {
-    border-radius: 0.5rem;
-    overflow: hidden;
-  }
-
-  @keyframes spin {
-    to {
-      transform: rotate(360deg);
-    }
-  }
-
-  .animate-spin {
-    animation: spin 1s linear infinite;
-  }
-
-  /* New pulse loader styles */
-  .pulse-loader {
-    display: flex;
-    align-items: center;
-    gap: 2px;
-    height: 20px;
-    width: 20px;
-    justify-content: center;
-    position: relative;
-  }
-
-  @keyframes pulse {
-    0%, 100% {
-      transform: scaleY(1);
-    }
-    50% {
-      transform: scaleY(1.35);
-    }
   }
 </style>
